+++ angular-app/src/app/app.component.ts
@@ -1,12 +1,15 @@
 import { Component } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterOutlet } from '@angular/router';
-import { PruebaComponent } from './prueba/prueba.component';
+
+import { HomeComponent } from './components/home/home.component';
+import { PruebaComponent } from './components/prueba/prueba.component';
+
 
 @Component({
   selector: 'app-root',
   standalone: true,
-  imports: [RouterOutlet, CommonModule, PruebaComponent],
+  imports: [RouterOutlet, CommonModule, HomeComponent, PruebaComponent],
   templateUrl: './app.component.html',
   styleUrl: './app.component.css'
 })
