+++ angular-app/src/app/app.routes.ts
@@ -1,8 +1,10 @@
 import { Routes } from '@angular/router';
 
-import { PruebaComponent } from './prueba/prueba.component';
+import { PruebaComponent } from './components/prueba/prueba.component';
+import { HomeComponent } from './components/home/home.component';
 
 
 export const routes: Routes = [
-  { path: 'prueba', component: PruebaComponent }
+    { path: 'home', component: HomeComponent},
+    { path: 'prueba', component: PruebaComponent }
 ];
