+++ angular-app/src/app/services/prueba/prueba.service.ts
@@ -1,7 +1,7 @@
 // prueba.service.ts
 import { Injectable } from '@angular/core';
 import { Observable, of } from 'rxjs';
-//import { HttpClient } from '@angular/common/http';
+import { HttpClient } from '@angular/common/http';
 
 @Injectable({
   providedIn: 'root'
@@ -9,11 +9,11 @@ import { Observable, of } from 'rxjs';
 export class PruebaService {
   private apiUrl = 'http://127.0.0.1:8000/sensors/last-sensor-entry/1';
 
-  //constructor(private http: HttpClient) { }
+  constructor(private http: HttpClient) { }
 
   obtenerEstado(): Observable<any> {
-    //return this.http.get<any>(this.apiUrl);
-    return of({
+    return this.http.get<any>(this.apiUrl);
+    /*return of({
       "battery": {
         "status-data": "87.9 %"
       },
@@ -31,7 +31,7 @@ export class PruebaService {
       }
     }
     );
-
+    */
   }
 }
